Add a Codebook Entry to a Data Frame Column

Function to update a column of a data frame to have a codebook entry.

abcddata_codebook_add_entry(
  dtf_base,
  chr_column,
  chr_description = "",
  lst_values_and_labels = NULL,
  lst_collected_over = NULL,
  chr_source_files = "",
  chr_source_variables = "",
  chr_notes = "",
  chr_category = "",
  chr_data_type = "",
  chr_reporter = ""
)

Arguments

dtf_base

A data frame.

chr_column

A character string, the column name of the variable for which a codebook entry will be added.

chr_description

A character string, a human-readable description of the variable and what it is for.

lst_values_and_labels

A list of two vectors, the values and their associated labels if applicable.

lst_collected_over

A list of two vectors giving the levels and associated variables over which the given variable was collected.

chr_source_files

A character vector, the source files from which the variable was taken.

chr_source_variables

A character vector, the original names of the source variables from which the variable was derived.

chr_notes

A character vector, additional notes on the variable.

chr_category

A character string, the category type for the variable (see abcddata_codebook_column_types).

chr_data_type

A character string, the data type for the variable (see abcddata_codebook_column_types).

chr_reporter

A character string, the responder type for the variable (see abcddata_codebook_column_types).

Value

A data frame with the associated column updated with a new attribute.

Examples

dtf_base <- data.frame(
  IDS.CHR.GD.Participant = LETTERS[1:3],
  QST.INT.CS.Example = 4:6
)

dtf_base <- abcddata_codebook_add_entry(
  dtf_base, 'IDS.CHR.GD.Participant',
  chr_description = 'Participant identifier'
)
abcddata_codebook_display( dtf_base$IDS.CHR.GD.Participant )
#> Variable: 
#>   IDS.CHR.GD.Participant
#> Category: 
#>   Identifiers
#> Data type: 
#>   Character string
#> Reporter: 
#>   General details
#> Description: 
#>   Participant identifier

dtf_base <- abcddata_codebook_add_entry(
  dtf_base, 'QST.INT.CS.Example',
  chr_description = 'Example questionnaire responses'
)
abcddata_codebook_display( dtf_base$QST.INT.CS.Example )
#> Variable: 
#>   QST.INT.CS.Example
#> Category: 
#>   Questionnaire
#> Data type: 
#>   Integer
#> Reporter: 
#>   Child reporting on self
#> Description: 
#>   Example questionnaire responses