Column Types for ABCD Codebook Entries

Function to convert abbreviations used in a column name into the associated label to include in a codebook entry for ABCD data.

abcddata_codebook_column_types(chr_column_name = "", chr_abbr = NULL)

Arguments

chr_column_name

A character string, the column name to process. If the empty character set '', lists the abbreviations and associated labels for each abbreviation type.

chr_abbr

A character string, the abbreviation type to consider, either 'Category', 'Data type', or 'Responder'.

Value

Either a character string, the label associated with the abbreviation of interest, or displays a message to the console listing abbreviations and labels.

Examples

abcddata_codebook_column_types()
#> Variable categories
#>   IDS = Identifiers
#>   DTQ = Data quality
#>   INC = Indices for inclusion
#>   SSS = Session details
#>   SMP = Sample characteristics
#>   PHS = Physical health
#>   MNT = Mental health
#>   IMG = Imaging data
#>   CGN = Cognitive functioning
#>   BMR = Biomarkers
#>   ENV = Environment
#>   SBS = Substance use
#>   QST = Questionnaire
#> 
#> Data types
#>   INT = Integer
#>   DBL = Double precision floating point number
#>   NMR = Numeric
#>   LGC = Logical
#>   CHR = Character string
#>   DTT = R posixt datetime variable
#>   FCT = Enumerated type - factor
#> 
#> Reporter types
#>   CS = Child reporting on self
#>   PC = Parents reporting on child
#>   PS = Parents reporting on self
#>   PO = Parents reporting on other
#>   GD = General details

chr_column <- 'IDS.CHR.GD.Subject'
abcddata_codebook_column_types( chr_column, 'Category' )
#>         Label 
#> "Identifiers" 
abcddata_codebook_column_types( chr_column, 'Data type' )
#>              Label 
#> "Character string" 
abcddata_codebook_column_types( chr_column, 'Reporter' )
#>             Label 
#> "General details"