Initialize Long-Form Data Frame

Function to initialize a long-form data frame for the ABCD® study with identifiers for participants, families, and sites, as well as session details and group splits for reproducible matched samples (Feczko et al., 2021).

abcddata_initialize_long_form(
  chr_files = c("abcd_y_lt.csv", "gen_y_pihat.csv"),
  chr_paths = c("core/abcd-general", "core/genetics"),
  lgc_progress = TRUE
)

Arguments

chr_files

A character vector with the file names for 1) the data on longitudinal tracking of the youths, and 2) genetics for the youths.

chr_paths

A character vector with the associated folder paths for the longitudinal tracking and genetics files.

lgc_progress

A logical value; if TRUE displays the progress of the data processing.

Value

A data frame.

Details

See https://abcdstudy.org/ for details on the ABCD® study, and refer to Saragosa-Harris et al. (2022) for helpful notes and instructions for researchers.

References

Feczko, E., Conan, G., Marek, S., Tervo-Clemmens, B., Cordova, M., Doyle, O., ... & Fair, D. A. (2021). Adolescent Brain Cognitive Development (ABCD) community MRI collection and utilities. BioRxiv. https://www.biorxiv.org/content/10.1101/2021.07.09.451638v1.abstract

Saragosa-Harris, N. M., Chaku, N., MacSweeney, N., Williamson, V. G., Scheuplein, M., Feola, B., ..., & Mills, K. L. (2022). A practical guide for researchers and reviewers using the ABCD Study and other large longitudinal datasets. Developmental Cognitive Neuroscience, 55, 1-11. https://doi.org/10.1016/j.dcn.2022.101115

Author

Kevin Potter